+++ src/test/java/im/aop/loggers/messageinterpolation/DefaultToStringStrategyFactoryTests.java
@@ -35,8 +35,8 @@ class DefaultToStringStrategyFactoryTests {
   static class TestToStringStrategy implements ToStringStrategy {
 
     @Override
-    public boolean supports(Class<?> type) {
-      return type == TestClass.class;
+    public boolean supports(Object object) {
+      return object instanceof TestClass;
     }
 
     @Override